# internal/features/user/adapters/http/handler.go
package userhttp

import (
	"net/http"

	userapp "github.com/caiohenrique/go-api-template/internal/features/user/app"
	"github.com/caiohenrique/go-api-template/internal/platform/ginbind"
	"github.com/caiohenrique/go-api-template/internal/platform/httperr"
	"github.com/caiohenrique/go-api-template/internal/platform/requestctx"
	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
)

// Handler exposes HTTP endpoints for users.
type Handler struct {
	svc *userapp.Service
}

// NewHandler creates the HTTP handler.
func NewHandler(svc *userapp.Service) *Handler {
	return &Handler{svc: svc}
}

// RegisterRoutes registers routes on the given router group.
func (h *Handler) RegisterRoutes(rg *gin.RouterGroup, authMiddleware gin.HandlerFunc) {
	rg.POST("/users", h.Create)
	rg.GET("/users/me", authMiddleware, h.Me)
}

// @Summary			Criar usuário
// @Description		Registra um novo usuário.
// @Tags			users
// @Accept			json
// @Produce			json
// @Param			body	body		userapp.CreateUserInput	true	"Dados do usuário"
// @Success			201		{object}	userapp.UserView
// @Failure			400		{object}	httperr.ErrorResponse
// @Failure			409		{object}	httperr.ErrorResponse
// @Failure			500		{object}	httperr.ErrorResponse
// @Router			/users [post]
func (h *Handler) Create(c *gin.Context) {
	in, ok := ginbind.JSON[userapp.CreateUserInput](c, "invalid json body")
	if !ok {
		return
	}

	out, err := h.svc.Create(c.Request.Context(), in)
	if err != nil {
		httperr.WriteServiceError(c, err)
		return
	}

	c.JSON(http.StatusCreated, out)
}

// @Summary			Usuário autenticado
// @Description		Retorna o perfil do usuário do token JWT.
// @Tags			users
// @Accept			json
// @Produce			json
// @Security		BearerAuth
// @Success			200	{object}	userapp.UserView
// @Failure			401	{object}	httperr.ErrorResponse
// @Failure			404	{object}	httperr.ErrorResponse
// @Failure			500	{object}	httperr.ErrorResponse
// @Router			/users/me [get]
func (h *Handler) Me(c *gin.Context) {
	id, ok := requestctx.UserID(c.Request.Context())
	if !ok || id == uuid.Nil {
		httperr.WriteError(c, http.StatusUnauthorized, "unauthorized")
		return
	}

	out, err := h.svc.GetByID(c.Request.Context(), id)
	if err != nil {
		httperr.WriteServiceError(c, err)
		return
	}

	c.JSON(http.StatusOK, out)
}